// dashboard/src/utils/hasura-api/generated/schemas/createRemoteSchemaRemoteRelationshipArgs.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Hasura Remote Schema API
 * API for managing remote schemas in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateRemoteSchemaRemoteRelationshipArgsDefinition } from './createRemoteSchemaRemoteRelationshipArgsDefinition';

export interface CreateRemoteSchemaRemoteRelationshipArgs {
  /** Name of the relationship. This will be used as the field name in the source type */
  name: string;
  /** Name of the source remote schema */
  remote_schema: string;
  /** Name of the type in the remote schema */
  type_name: string;
  /** Relationship definition */
  definition: CreateRemoteSchemaRemoteRelationshipArgsDefinition;
}